requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.23.0
pillow>=10.0.0
pyyaml>=6.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
paho-mqtt>=1.6.1
pymodbus>=3.5.4
streamlit>=1.33.0
